feat: Retry on podcast processing error.
Added new RecoverableError that is thrown every time there is a recoverable error, that is, the message can be retried later.
The retry mecanism is done using a dlq queue, which itselft uses the main queue as dlq with a set message ttl. This way, I dont need to control the interval between retries manually.
The retry limit is controlled using the "x-death" message header.